Remane ConsolidatedState to ConsolidatedObservation
Also rename Object in ConsolidatedObservation and Observation to State.
This is to better describe that the tt:Object structure represents the
state of the object that is valid at specific time in the plain
observation case and the state of the object that is valid over a period
in the consolidated observation case.