Add image crop box
The goal is to add a "crop" box somewhere to an object image that is
related to the full camera image. It might be of interest for a user to know
the part of the full image the object image came from. It is not always
possible to use the object detection bounding box as it migt be different to
the image crop box. One example is when padding around the object detection
bounding box is added in the image makeing the image crop box bigger
than the object detection bounding box.